refactor: 컴포넌트 폴더 구조 분리 및 변경
