+++ src/components/ShoppingCart/CartItem/CartItem.style.tsx
