+++ src/components/ShoppingCart/CartItem/CartItem.tsx
