+++ src/components/ShoppingCart/TotalPriceContainer/TotalPriceContainer.style.tsx
