+++ src/pages/ShoppingCartPage/ShoppingCartPage.tsx
@@ -2,8 +2,8 @@ import { useLoaderData, Await, Link } from 'react-router-dom';
 import Header, { MainLogoButton } from '../../components/Header/Header';
 import TitleContainer from '../../components/common/TitleContainer/TitleContainer';
 import SubmitButton from '../../components/common/SubmitButton/SubmitButton';
-import CartItemList from '../../components/ShoppingCartPage/CartItemList/CartItemList';
-import TotalPriceContainer from '../../components/common/TotalPriceContainer/TotalPriceContainer';
+import CartItemList from '../../components/ShoppingCart/CartItemList/CartItemList';
+import TotalPriceContainer from '../../components/ShoppingCart/TotalPriceContainer/TotalPriceContainer';
 import type { TCartItem } from '../../types/CartItem.type';
 import { EmptyCart } from '../../assets';
 import { PATHS } from '../../constants/PATHS';
