+++ src/recoil/Coupon/atoms/atoms.test.ts
@@ -1,10 +1,11 @@
 import { renderHook, act } from '@testing-library/react';
 import { RecoilRoot, useRecoilState } from 'recoil';
-import { selectedCouponIdListState, totalDiscountPriceState } from './atoms';
+import { selectedCouponListState } from './atoms';
+import { mockCoupons } from '../../../mocks/coupons';
 
-describe('selectedCouponIdListState', () => {
+describe('selectedCouponListState', () => {
   it('초기 배열은 빈 배열이다.', () => {
-    const { result } = renderHook(() => useRecoilState(selectedCouponIdListState), {
+    const { result } = renderHook(() => useRecoilState(selectedCouponListState), {
       wrapper: RecoilRoot,
     });
 
@@ -12,32 +13,12 @@ describe('selectedCouponIdListState', () => {
   });
 
   it('배열에 데이터를 삽입할 수 있다.', () => {
-    const { result } = renderHook(() => useRecoilState(selectedCouponIdListState), {
+    const { result } = renderHook(() => useRecoilState(selectedCouponListState), {
       wrapper: RecoilRoot,
     });
 
-    act(() => result.current[1]([1, 2, 3, 4]));
+    act(() => result.current[1](mockCoupons));
 
     expect(result.current[0].length).toBe(4);
-  });
-});
-
-describe('totalDiscountPriceState', () => {
-  it('초기 상태는 0 이다.', () => {
-    const { result } = renderHook(() => useRecoilState(totalDiscountPriceState), {
-      wrapper: RecoilRoot,
-    });
-
-    expect(result.current[0]).toBe(0);
-  });
-
-  it('상태를 변화시킬 수 있다.', () => {
-    const { result } = renderHook(() => useRecoilState(totalDiscountPriceState), {
-      wrapper: RecoilRoot,
-    });
-
-    act(() => result.current[1](1000));
-
-    expect(result.current[0]).toBe(1000);
   });
 });
