+++ src/utils/time.ts
@@ -1,7 +1,7 @@
 const KOREAN_TIME_INTERVAL = 9 * 60 * 60 * 1000;
 
 export const getKoreanTime = () => {
-  const current = new Date(2024, 3, 20, 4, 30, 30);
+  const current = new Date();
   const utc = current.getTime() + current.getTimezoneOffset() * 60 * 1000;
   return new Date(utc + KOREAN_TIME_INTERVAL);
 };
